requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
boto3>=1.34.0
opensearch-py>=2.4.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.5.0
neo4j>=5.14.0
streamlit>=1.28.0
botocore>=1.34.0 